Add parsing of an additional, local config
Tested with `dixit/config.local.json` as:
```
{
    // Port for the server to listen on.
    "port" : 8889,

    "limits" : {
        "min_clue_length": 1,
        "min_user_name" : 2
    }
}
```

All three settings were overridden as expected.

The main motivation was exactly these 3: in some languages names can be
shorter than 3 char, and clues can be meaningful even with less than 3 chars.
And of course it's nice being able to override the port without having a
change in the repo.