Add support for hiding games
Not deleting them as that changes indexes, but allow the host to hide
them so the list doesn't get cluttered with all the test games :)

This is another variation on https://github.com/arvoelke/Dixit/pull/5
The main difference is that the game table is nowconstructed with
jQuery, which allows for an easy way to attach the game's name on the
relevant row. This in turn removes the need for storing game data.